# apps/api/src/main/resources/static/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API仕様書",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "ローカル環境"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "書籍情報を検索して取得する",
        "operationId": "searchBooks",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "description": "ISBN-13:完全一致",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 9784873115658
          },
          {
            "name": "title",
            "in": "query",
            "description": "書籍タイトル:部分一致",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priceFrom",
            "in": "query",
            "description": "価格:下限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "priceTo",
            "in": "query",
            "description": "価格:上限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "ステータス",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "UNPUBLISHED",
                  "PUBLISHED",
                  "OUT_OF_PRINT"
                ]
              }
            }
          },
          {
            "name": "publishedAtStart",
            "in": "query",
            "description": "出版日時:ISO8601:開始",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "publishedAtEnd",
            "in": "query",
            "description": "出版日時:ISO8601:終了",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "order",
            "in": "query",
            "description": "並び順",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-published_at",
              "enum": [
                "+price",
                "-price",
                "+published_at",
                "-published_at"
              ]
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "取得位置",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "取得数",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchBooksResponse"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "description": "新しい書籍情報を作成する",
        "operationId": "createBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "指定したIDの書籍情報を取得する",
        "operationId": "findBookById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindBookByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "description": "指定したIDの書籍情報を削除する",
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Book"
        ],
        "description": "指定したIDの書籍情報を更新する",
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/books/download": {
      "get": {
        "tags": [
          "Book"
        ],
        "description": "[Internal Use Only]書籍情報をCSV形式でダウンロードする",
        "operationId": "downloadBooks",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "description": "ISBN-13:完全一致",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 9784873115658
          },
          {
            "name": "title",
            "in": "query",
            "description": "書籍タイトル:部分一致",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priceFrom",
            "in": "query",
            "description": "価格:下限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "priceTo",
            "in": "query",
            "description": "価格:上限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "ステータス",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "UNPUBLISHED",
                  "PUBLISHED",
                  "OUT_OF_PRINT"
                ]
              }
            }
          },
          {
            "name": "publishedAtStart",
            "in": "query",
            "description": "出版日時:ISO8601:開始",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "publishedAtEnd",
            "in": "query",
            "description": "出版日時:ISO8601:終了",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "order",
            "in": "query",
            "description": "並び順",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-published_at",
              "enum": [
                "+price",
                "-price",
                "+published_at",
                "-published_at"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateBookBody": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": "9784873115658",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "description": "書籍タイトル",
            "example": "リーダブルコード",
            "minLength": 1
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          }
        },
        "required": [
          "authorId",
          "isbn",
          "publisherId",
          "status",
          "title"
        ]
      },
      "ValidationErrorResponse": {
        "type": "object",
        "description": "バリデーションエラーの詳細",
        "properties": {
          "field": {
            "type": "string",
            "description": "エラーが発生したフィールド",
            "example": "isbn"
          },
          "message": {
            "type": "string",
            "description": "エラーメッセージ",
            "example": "空白は許可されていません"
          }
        }
      },
      "ProblemDetailResponse": {
        "type": "object",
        "description": "Problem Details for HTTP APIs",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "description": "問題種別を識別するURI",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "問題の短い説明",
            "example": "Bad Request"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTPステータスコード",
            "example": 400
          },
          "detail": {
            "type": "string",
            "description": "問題の詳細説明",
            "example": "リクエストボディの検証に失敗しました"
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "description": "問題が発生したリソースのURI",
            "example": "/books"
          },
          "errorCode": {
            "type": "string",
            "description": "アプリケーション固有のエラーコード",
            "example": "dw1001"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
      "UpdateBookBody": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": "9784873115658"
          },
          "title": {
            "type": "string",
            "description": "書籍タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          }
        }
      },
      "ResponsePagination": {
        "type": "object",
        "properties": {
          "returnedCount": {
            "type": "integer",
            "format": "int32",
            "description": "取得件数",
            "example": 0
          },
          "currentPage": {
            "type": "integer",
            "format": "int32",
            "description": "現在のページ番号",
            "example": 100
          },
          "totalCount": {
            "type": "integer",
            "format": "int32",
            "description": "総件数",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "description": "総ページ数",
            "example": 1
          }
        },
        "required": [
          "currentPage",
          "returnedCount",
          "totalCount",
          "totalPages"
        ]
      },
      "SearchBooksResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchBooksResponseItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/ResponsePagination"
          }
        }
      },
      "SearchBooksResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "書籍ID",
            "example": 1
          },
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": "9784873115658"
          },
          "title": {
            "type": "string",
            "description": "タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "authorName": {
            "type": "string",
            "description": "著者名",
            "example": "Dustin Boswell"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          },
          "publisherName": {
            "type": "string",
            "description": "出版社名",
            "example": "O\u0027Reilly"
          }
        },
        "required": [
          "id",
          "isbn",
          "status",
          "title"
        ]
      },
      "FindBookByIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "書籍ID",
            "example": 1
          },
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": "9784873115658"
          },
          "title": {
            "type": "string",
            "description": "タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "authorName": {
            "type": "string",
            "description": "著者名",
            "example": "Dustin Boswell"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          },
          "publisherName": {
            "type": "string",
            "description": "出版社名",
            "example": "O\u0027Reilly"
          }
        },
        "required": [
          "id",
          "isbn",
          "status",
          "title"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "ApiKeyAuth",
        "in": "header"
      }
    }
  }
}